GOpamine/backend/requirements.txt
# --- Flask Framework & Tiện ích ---
Flask
Flask-CORS
Flask-Login
Flask-WTF
WTForms
python-dotenv
psycopg2
psycopg2-binary

# --- AI (Google Gemini) ---
google-generativeai
nltk

# --- Bản đồ & Thuật toán (cho astar.py và map_trans) ---
folium
geopy
networkx
osmnx
# pandas thường đi kèm để xử lý dữ liệu bản đồ/Excel
pandas

# --- Cơ sở dữ liệu & Bảo mật (cho auth.py) ---
bcrypt
PyJWT

# --- HTTP Client ---
requests

# --- Các thư viện tiện ích khác ---
gunicorn
